# backend/docs/swagger.yaml
definitions:
  consts.SourceType:
    enum:
    - dingtalk
    - feishu
    - wecom
    - oauth
    - cas
    - ldap
    type: string
    x-enum-varnames:
    - SourceTypeDingTalk
    - SourceTypeFeishu
    - SourceTypeWeCom
    - SourceTypeOAuth
    - SourceTypeCAS
    - SourceTypeLDAP
...
  domain.AuthGetResp:
    properties:
      auth_type:
        $ref: '#/definitions/domain.AuthType'
      source_type:
        $ref: '#/definitions/consts.SourceType'
    type: object
  domain.AuthLoginSimpleReq:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  domain.AuthType:
    enum:
    - ""
    - simple
    - enterprise
    type: string
    x-enum-comments:
      AuthTypeEnterprise: 企业认证
      AuthTypeNull: 无认证
      AuthTypeSimple: 简单口令
    x-enum-descriptions:
    - 无认证
    - 简单口令
    - 企业认证
    x-enum-varnames:
    - AuthTypeNull
    - AuthTypeSimple
    - AuthTypeEnterprise
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /share/v1/auth/get:
    get:
      consumes:
      - application/json
      description: AuthGet
      operationId: v1-AuthGet
      parameters:
      - description: kb_id
        in: header
        name: X-KB-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.AuthGetResp'
              type: object
      summary: AuthGet
      tags:
      - share_auth
  /share/v1/auth/login/simple:
    post:
      consumes:
      - application/json
      description: AuthLoginSimple
      operationId: v1-AuthLoginSimple
      parameters:
      - description: kb_id
        in: header
        name: X-KB-ID
        required: true
        type: string
      - description: para
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/domain.AuthLoginSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: AuthLoginSimple
      tags:
      - share_auth
...
swagger: "2.0"